refactor: move every validation shape validator to a separate file
Refs: #SFB-39
