+++ app/utils/validation-shape-validators.js
@@ -1,19 +1,31 @@
 import { Statement } from 'rdflib';
-import { FORM, RDF } from './rdflib';
+import { FORM } from './rdflib';
 import { defaultCountryCode } from '../components/rdf-form-fields/country-code-concept-scheme-selector';
+import { isMaxCharacterValueAddedToMaxLengthValidation } from './validation-shape-validators/is-max-character-value-added';
+import { isDefaultCountryCodeAddedToValidPhoneNumber } from './validation-shape-validators/is-country-code-added';
+import { isExactValueAddedToExactValueConstraint } from './validation-shape-validators/is-exact-value-added';
+import { isRdfTypeInSubjects } from './validation-shape-validators/is-rdf-type-in-subjects';
 
 export function areValidationsInGraphValidated(store, graph) {
-  const validationNodes = getValidationNodesInGraph(store, graph);
+  const validationNodes = getAllValidationNodesInGraph(store, graph);
+  const defaultCountryCodesAdded = isDefaultCountryCodeAddedToValidPhoneNumber(
+    store,
+    graph
+  );
+  if (!defaultCountryCodesAdded.isAdded) {
+    for (const subject of defaultCountryCodesAdded.subjectsToAddDefaultTo) {
+      insertDefaultCountryCode(subject, store, graph);
+    }
+  }
 
   return ![
-    isRdfTypeInTriplesOfSubjects(validationNodes, store, graph),
+    isRdfTypeInSubjects(validationNodes, store, graph),
     isMaxCharacterValueAddedToMaxLengthValidation(store, graph),
     isExactValueAddedToExactValueConstraint(store, graph),
-    isCountryCodeAddedToValidPhoneNumber(store, graph),
   ].includes(false);
 }
 
-function getValidationNodesInGraph(store, graph) {
+function getAllValidationNodesInGraph(store, graph) {
   const subjectThatHaveValidations = store.match(
     undefined,
     FORM('validations'),
@@ -21,84 +33,6 @@ function getValidationNodesInGraph(store, graph) {
     graph
   );
   return subjectThatHaveValidations.map((statement) => statement.object);
-}
-
-function isRdfTypeInTriplesOfSubjects(subjects, store, graph) {
-  for (const subject of subjects) {
-    const typeMatches = store.match(subject, RDF('type'), undefined, graph);
-    if (!typeMatches.length >= 1) {
-      return false;
-    }
-  }
-
-  return true;
-}
-
-function isMaxCharacterValueAddedToMaxLengthValidation(store, graph) {
-  const maxLengthValidationSubjects = store.match(
-    undefined,
-    RDF('type'),
-    FORM('MaxLength'),
-    graph
-  );
-  for (const triple of maxLengthValidationSubjects) {
-    const maxCharactersValues = store.match(
-      triple.subject,
-      FORM('max'),
-      undefined,
-      graph
-    );
-    if (!maxCharactersValues.length >= 1) {
-      return false;
-    }
-  }
-
-  return true;
-}
-
-function isExactValueAddedToExactValueConstraint(store, graph) {
-  const exactValueConstraint = store.match(
-    undefined,
-    RDF('type'),
-    FORM('ExactValueConstraint'),
-    graph
-  );
-  for (const triple of exactValueConstraint) {
-    const exactValueValues = store.match(
-      triple.subject,
-      FORM('customValue'),
-      undefined,
-      graph
-    );
-    if (!exactValueValues.length >= 1) {
-      return false;
-    }
-  }
-
-  return true;
-}
-
-function isCountryCodeAddedToValidPhoneNumber(store, graph) {
-  const validPhoneNumber = store.match(
-    undefined,
-    RDF('type'),
-    FORM('ValidPhoneNumber'),
-    graph
-  );
-  for (const triple of validPhoneNumber) {
-    const countryCodeValues = store.match(
-      triple.subject,
-      FORM('defaultCountry'),
-      undefined,
-      graph
-    );
-
-    if (!countryCodeValues.length >= 1) {
-      insertDefaultCountryCode(triple.subject, store, graph);
-    }
-  }
-
-  return true;
 }
 
 function insertDefaultCountryCode(subject, store, graph) {
