+++ app/utils/validation-shape-validators/is-max-character-value-added.js
@@ -1,0 +1,23 @@
+import { FORM, RDF } from '../rdflib';
+
+export function isMaxCharacterValueAddedToMaxLengthValidation(store, graph) {
+  const maxLengthValidationSubjects = store.match(
+    undefined,
+    RDF('type'),
+    FORM('MaxLength'),
+    graph
+  );
+  for (const triple of maxLengthValidationSubjects) {
+    const maxCharactersValues = store.match(
+      triple.subject,
+      FORM('max'),
+      undefined,
+      graph
+    );
+    if (!maxCharactersValues.length >= 1) {
+      return false;
+    }
+  }
+
+  return true;
+}
