+++ app/utils/validation-shape-validators/is-rdf-type-in-subjects.js
@@ -1,0 +1,12 @@
+import { RDF } from '../rdflib';
+
+export function isRdfTypeInSubjects(subjects, store, graph) {
+  for (const subject of subjects) {
+    const typeMatches = store.match(subject, RDF('type'), undefined, graph);
+    if (!typeMatches.length >= 1) {
+      return false;
+    }
+  }
+
+  return true;
+}
